requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
streamlit==1.25.0
pandas==2.2.2
catboost==1.2.8
scikit-learn==1.3.2
pytest==7.4.2
requests==2.31.0
mlflow==2.14.1
python-dotenv==1.0.1  
